Delay patching some methods to avoid multithread issues.
