+++ Source/ArtilleryCompat/ArtilleryCompat/Harmony/Harmony_ArtilleryComp.cs
@@ -34,32 +34,6 @@ namespace CombatExtended.Compatibility.Artillery
         }
     }*/
 
-    // TODO: Remove this dirty hack to work around VFE bug
-    [HarmonyPatch(typeof(WorldArtilleryTracker), nameof(WorldArtilleryTracker.ExposeData))]
-    public class Harmony_WorldArtilleryTracker_ExposeData {
-	private static HashSet<ArtilleryComp> cached_cachedArtilleryCompsBombarding = null;
-	public static bool Prefix(WorldArtilleryTracker __instance) {
-	    if (Scribe.mode == LoadSaveMode.Saving) {
-		if (__instance.cachedArtilleryCompsBombarding.Any()) {
-		    cached_cachedArtilleryCompsBombarding = __instance.cachedArtilleryCompsBombarding;
-		    __instance.cachedArtilleryCompsBombarding = new HashSet<ArtilleryComp>();
-		}
-		else {
-		    cached_cachedArtilleryCompsBombarding = null;
-		}
-	    }
-	    return true;
-	}
-	public static void Postfix(WorldArtilleryTracker __instance) {
-	    if (Scribe.mode == LoadSaveMode.Saving) {
-		if (cached_cachedArtilleryCompsBombarding != null) {
-		    __instance.cachedArtilleryCompsBombarding = cached_cachedArtilleryCompsBombarding;
-		    cached_cachedArtilleryCompsBombarding = null;
-		}
-	    }
-	}
-    }
-
     [HarmonyPatch(typeof(ArtilleryComp), "BombardmentTick")]
     public class Harmony_ArtilleryComp_BombardmentTick {
         public static IEnumerable<CodeInstruction> Transpiler(ILGenerator gen, IEnumerable<CodeInstruction> instructions)
